move graphql types into graphql folder (#8420)
