+++ backend/src/graphql/types/enum/EmailNotificationSettingsName.gql
