+++ backend/src/graphql/types/enum/EmailNotificationSettingsType.gql
