+++ backend/src/graphql/types/enum/Emotion.gql
