+++ backend/src/graphql/types/enum/GroupActionRadius.gql
