+++ backend/src/graphql/types/enum/GroupMemberRole.gql
