+++ backend/src/graphql/types/enum/GroupType.gql
