+++ backend/src/graphql/types/enum/PostType.gql
