+++ backend/src/graphql/types/enum/ShoutTypeEnum.gql
@@ -1,0 +1,3 @@
+enum ShoutTypeEnum {
+  Post
+}
