+++ backend/src/graphql/types/type/Badge.gql
