+++ backend/src/graphql/types/type/Category.gql
