+++ backend/src/graphql/types/type/Group.gql
