+++ backend/src/graphql/types/type/Image.gql
