+++ backend/src/graphql/types/type/InviteCode.gql
