+++ backend/src/graphql/types/type/NOTIFIED.gql
