+++ backend/src/graphql/types/type/Reward.gql.unused
@@ -1,16 +1,7 @@
-enum ShoutTypeEnum {
-  Post
-}
-
 type Reward {
   id: ID!
   user: User @relation(name: "REWARDED", direction: "IN")
   rewarderId: ID
   createdAt: String
   badge: Badge @relation(name: "REWARDED", direction: "OUT")
-}
-
-type SharedInboxEndpoint {
-  id: ID!
-  uri: String
 }
