+++ backend/src/graphql/types/type/Search.gql
