+++ backend/src/graphql/types/type/SharedInboxEndpoint.gql.old
@@ -1,0 +1,4 @@
+type SharedInboxEndpoint {
+  id: ID!
+  uri: String
+}
