+++ backend/src/graphql/types/type/SocialMedia.gql
