+++ backend/src/graphql/types/type/Statistics.gql
